remove references to API constants, because it breaks tests of the UI
