+++ redisinsight/api/src/constants/telemetry-events.ts
@@ -104,13 +104,4 @@ export enum CommandType {
   Module = 'module',
 }
 
-export const ToggleAnalyticsReason = {
-  None: 'none',
-  OauthAgreement: 'oauth-agreement',
-  Google: 'google',
-  GitHub: 'github',
-  SSO: 'sso',
-  User: 'user',
-} as const;
-export type ToggleAnalyticsReasonType = typeof ToggleAnalyticsReason[keyof typeof ToggleAnalyticsReason];
 export const unknownCommand = 'unknown';
