+++ redisinsight/api/src/modules/settings/constants/settings.ts
@@ -1,0 +1,9 @@
+export const ToggleAnalyticsReason = {
+  None: 'none',
+  OauthAgreement: 'oauth-agreement',
+  Google: 'google',
+  GitHub: 'github',
+  SSO: 'sso',
+  User: 'user',
+} as const;
+export type ToggleAnalyticsReasonType = typeof ToggleAnalyticsReason[keyof typeof ToggleAnalyticsReason];
