+++ redisinsight/api/src/modules/settings/dto/settings.dto.ts
@@ -15,8 +15,9 @@ import { pickDefinedAgreements } from 'src/dto/dto-transformer';
 import { Default } from 'src/common/decorators';
 import config from 'src/utils/config';
 import { IAgreementSpec } from 'src/modules/settings/models/agreements.interface';
-import { ToggleAnalyticsReason, ToggleAnalyticsReasonType } from 'src/constants/telemetry-events';
+import { ToggleAnalyticsReason, ToggleAnalyticsReasonType } from 'src/modules/settings/constants/settings';
 
+export { ToggleAnalyticsReason, ToggleAnalyticsReasonType };
 const REDIS_SCAN_CONFIG = config.get('redis_scan');
 const WORKBENCH_CONFIG = config.get('workbench');
