+++ redisinsight/ui/src/components/consents-settings/ConsentsSettings.tsx
@@ -22,7 +22,6 @@ import cx from 'classnames'
 import { compareConsents } from 'uiSrc/utils'
 import { updateUserConfigSettingsAction, userSettingsSelector } from 'uiSrc/slices/user/user-settings'
 import { sendEventTelemetry, TelemetryEvent } from 'uiSrc/telemetry'
-import { ToggleAnalyticsReason } from 'apiSrc/constants/telemetry-events'
 import ConsentOption from './ConsentOption'
 
 import styles from './styles.module.scss'
@@ -190,7 +189,7 @@ const ConsentsSettings = ({ onSubmitted }: Props) => {
     }
     const settings: Record<string, any> = { agreements: values }
     if (values.analytics) {
-      settings.analyticsReason = ToggleAnalyticsReason.User
+      settings.analyticsReason = 'user'
     }
     dispatch(updateUserConfigSettingsAction(settings, onSubmitted))
   }
